feedback for resit implemented 0.1

+++ GymApplication/WorkoutLogic/CardioWorkout.cs
@@ -9,10 +9,10 @@ namespace GymApplication.WorkoutLogic
     public class CardioWorkout : IWorkout
     {
         public string Name { get; private set; }
-        public int Intensity { get; private set; }
+        public double Intensity { get; private set; }
         public List<Exercise> Exercises { get; private set; }
 
-        public CardioWorkout(string name, int intensity, List<Exercise> exercises)
+        public CardioWorkout(string name, double intensity, List<Exercise> exercises)
         {
             Name = name;
             Intensity = intensity;
