+++ GymApplication/WorkoutLogic/FlexibilityWorkout.cs
@@ -9,10 +9,10 @@ namespace GymApplication.WorkoutLogic
     public class FlexibilityWorkout : IWorkout
     {
         public string Name { get; private set; }
-        public int Intensity { get; private set; }
+        public double Intensity { get; private set; }
         public List<Exercise> Exercises { get; private set; }
 
-        public FlexibilityWorkout(string name, int intensity, List<Exercise> exercises)
+        public FlexibilityWorkout(string name, double intensity, List<Exercise> exercises)
         {
             Name = name;
             Intensity = intensity;
