+++ GymApplication/WorkoutLogic/StrengthWorkout.cs
@@ -9,10 +9,10 @@ namespace GymApplication.WorkoutLogic
     public class StrengthWorkout : IWorkout
     {
         public string Name { get; private set; }
-        public int Intensity { get; private set; }
+        public double Intensity { get; private set; }
         public List<Exercise> Exercises { get; private set; }
 
-        public StrengthWorkout(string name, int intensity, List<Exercise> exercises)
+        public StrengthWorkout(string name, double intensity, List<Exercise> exercises)
         {
             Name = name;
             Intensity = intensity;
